+++ contact/association/delete/index.js
@@ -1,19 +1,26 @@
 const {getSingular} = require("../../../utils/crm.types")
 
 const onInput = (node, config) => async (msg, send, done) => {
+  const hubspot = await node.getHubspotWrapper()
+
+  if (!hubspot) {
+    node.error("hubspot account missing");
+    return
+  }
+
   if (
-    !msg[config.inputDealId]
+    !msg[config.inputContactId]
     || !msg[config.inputLinkId]
   ) {
-    node.error(`msg.${config.inputDealId} must be set`);
+    node.error(`msg.${config.inputContactId} must be set`);
     node.error(`msg.${config.inputLinkId} must be set`);
 
     return
   }
 
   try {
-    msg[String(config.output || 'payload')] = (await node.hubspot.crm.contacts.associationsApi.archive(
-      msg[String(config.inputDealId)],
+    msg[String(config.output || 'payload')] = (await hubspot.crm.contacts.associationsApi.archive(
+      msg[String(config.inputContactId)],
       config.inputLinkType,
       msg[String(config.inputLinkId)],
       `contact_to_${getSingular(config.inputLinkType)}`
@@ -34,14 +41,10 @@ module.exports = (RED) => {
   function node(config) {
     RED.nodes.createNode(this, config)
 
-    this.hubspot = RED.nodes.getNode(config.account)?.hubspot
-
-    if (!this.hubspot) {
-      return
-    }
+    this.getHubspotWrapper = () => RED.nodes.getNode(config.account)?.getHubspotWrapper()
 
     this.on('input', onInput(this, config))
   }
 
-  RED.nodes.registerType("hubspot-deal-association-delete", node);
+  RED.nodes.registerType("hubspot-contact-association-delete", node);
 }