+++ deal/create/index.js
@@ -1,7 +1,14 @@
 
 const onInput = (node, config) => async (msg, send, done) => {
+  const hubspot = await node.getHubspotWrapper()
+
+  if (!hubspot) {
+    node.error("hubspot account missing");
+    return
+  }
+
   try {
-    msg[String(config.output || 'payload')] = (await node.hubspot.crm.deals.basicApi.create(msg[config.inputData])).body
+    msg[String(config.output || 'payload')] = (await hubspot.crm.deals.basicApi.create(msg[config.inputData])).body
 
     send(msg)
   } catch (e) {
@@ -17,11 +24,7 @@ module.exports = (RED) => {
   function node(config) {
     RED.nodes.createNode(this, config)
 
-    this.hubspot = RED.nodes.getNode(config.account)?.hubspot
-
-    if (!this.hubspot) {
-      return
-    }
+    this.getHubspotWrapper = () => RED.nodes.getNode(config.account)?.getHubspotWrapper()
 
     this.on('input', onInput(this, config))
   }