+++ deal/update/index.js
@@ -1,5 +1,12 @@
 
 const onInput = (node, config) => async (msg, send, done) => {
+  const hubspot = await node.getHubspotWrapper()
+
+  if (!hubspot) {
+    node.error("hubspot account missing");
+    return
+  }
+
   if (!['number', 'string'].includes(typeof msg[config.inputId])) {
     node.error(`msg.${config.inputId} should be either string or number`);
 
@@ -7,7 +14,7 @@ const onInput = (node, config) => async (msg, send, done) => {
   }
 
   try {
-    msg[String(config.output || 'payload')] = (await node.hubspot.crm.deals.basicApi.update(
+    msg[String(config.output || 'payload')] = (await hubspot.crm.deals.basicApi.update(
       msg[config.inputId],
       msg[config.inputData]
     )).body
@@ -27,11 +34,7 @@ module.exports = (RED) => {
   function node(config) {
     RED.nodes.createNode(this, config)
 
-    this.hubspot = RED.nodes.getNode(config.account)?.hubspot
-
-    if (!this.hubspot) {
-      return
-    }
+    this.getHubspotWrapper = () => RED.nodes.getNode(config.account)?.getHubspotWrapper()
 
     this.on('input', onInput(this, config))
   }